refactor(core): break core/_internal → harness reverse imports
The kernel's constitution layer (``core/_internal``) was reverse-importing
``harness`` in three places, inverting the layer rule documented in
``self-modifiable-architecture.md``: harness sits *above* core, not below.

Two reverse imports eliminated outright:

* ``core/_internal/compaction/compaction.py`` and
  ``core/_internal/compaction/branch_summarization.py`` imported
  ``SessionEntry`` (and ``SessionManager``) from
  ``harness/session_manager``. ``SessionEntry`` and friends are pure
  data shapes (frozen dataclasses, no I/O) and belong in ``core/abi``.
  Promoted to a new ``core/abi/session.py`` module exposing the data
  classes, helper constructors, and a minimal ``SessionTree`` Protocol
  (read-only ``get_branch`` / ``get_entry``). The full ``SessionManager``
  (persistence, fork/navigate) stays in ``harness`` and implements
  ``SessionTree`` implicitly. ``branch_summarization.collect_entries``
  now types its parameter as ``SessionTree``, so compaction code never
  reaches into the harness.

* ``ResourceWriter`` Protocol + ``WriteResult`` / ``BatchHandle`` /
  ``WriterAuthor`` / ``PathClass`` moved from
  ``harness/resource_writer.py`` to a new ``core/abi/resource.py``.
  ``harness/resource_writer.py`` re-exports them; the concrete
  ``GitBackedResourceWriter`` stays put.
  ``core/_internal/catalog/freeze.py`` now types against the Protocol
  and accepts an optional ``writer`` parameter. The remaining
  function-local ``from agentm.harness.resource_writer import
  GitBackedResourceWriter`` is the documented "default-impl escape
  hatch" for the catalog-freeze CLI / unit-test path: callers may
  still inject their own writer to bypass it.

The public ``harness/__init__`` re-exports of ``SessionEntry`` etc. are
unchanged, so existing extension code keeps working.
